Changed Integers to Int64s

+++ src/BzoTypes.hs
@@ -1,4 +1,5 @@
 module BzoTypes where
+import Data.Int
 import qualified Data.Text as T
 import Data.Map.Strict hiding (map)
 import HigherOrder
@@ -852,12 +853,12 @@ instance Show ExprModel where show = showExprModel
 
 data SymbolTable
   = SymbolTable {
-      st_iids   :: !(Map T.Text [(Integer, Integer)]),   -- To look up if an Identifier exists, and if so, what are the Table Indices(snd) and Files(fst)?
-      st_fids   :: !(Map T.Text Integer),                -- To look up if a File Identifier exists, and if so, what is the Table Index?
-      st_itable :: !(Map Integer (T.Text, Integer)),     -- For a given Table Index, what is the associated Identifier, and the file where it's defined?
-      st_ftable :: !(Map Integer T.Text),                -- For a given Table Index, what is the associated File Identifier?
-      st_itop   :: !Integer,                           -- What is the highest used Identifier Table Index?
-      st_ftop   :: !Integer }                          -- What is the highest used File Table Index?
+      st_iids   :: !(Map T.Text [(Int64, Int64)]),   -- To look up if an Identifier exists, and if so, what are the Table Indices(snd) and Files(fst)?
+      st_fids   :: !(Map T.Text Int64),              -- To look up if a File Identifier exists, and if so, what is the Table Index?
+      st_itable :: !(Map Int64 (T.Text, Int64)),     -- For a given Table Index, what is the associated Identifier, and the file where it's defined?
+      st_ftable :: !(Map Int64 T.Text),              -- For a given Table Index, what is the associated File Identifier?
+      st_itop   :: !Int64,                           -- What is the highest used Identifier Table Index?
+      st_ftop   :: !Int64 }                          -- What is the highest used File Table Index?
   deriving Show
